Refactor db connection code
